Better clean up

+++ headers/responsive_analog_read.h
@@ -1,6 +1,6 @@
 #ifndef RESPONSIVE_ANALOG_READ_H_
 #define RESPONSIVE_ANALOG_READ_H_
 
-float responsive_analog_read(float newValue);
+int responsive_analog_read(float newValue);
 
 #endif /* RESPONSIVE_ANALOG_READ_H_ */
